[identity] reintroduce check to make sure only wallet users can call deleteWalletUser RPC
Summary:
Depends on D12326

Devs will have to pull changes from D12326 to delete password accounts after this change is added to staging

Test Plan: i got a permission denied error trying to call deleteWalletUser from a password account

Reviewers: bartek, will

Reviewed By: will

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12335

+++ services/identity/src/grpc_services/authenticated.rs
@@ -376,6 +376,16 @@ impl IdentityClientService for AuthenticatedService {
 
     debug!("Attempting to delete wallet user: {}", user_id);
 
+    let maybe_username_and_password_file = self
+      .db_client
+      .get_username_and_password_file(&user_id)
+      .await
+      .map_err(handle_db_error)?;
+
+    if maybe_username_and_password_file.is_some() {
+      return Err(tonic::Status::permission_denied("password user"));
+    }
+
     self
       .db_client
       .delete_user(user_id)
